+++ R/TerminatorStagnationBatch.R
@@ -36,8 +36,8 @@ TerminatorStagnationBatch = R6Class("TerminatorStagnationBatch",
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
       param_set = ps(
-        n = p_int(lower = 1L, default = 1, tags = "required"),
-        threshold = p_dbl(lower = 0, default = 0, tags = "required")
+        n = p_int(lower = 1L, tags = "required"),
+        threshold = p_dbl(lower = 0, tags = "required")
       )
       param_set$values = list(n = 1, threshold = 0)