+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * fix: `OptimizerIrace` failed with logical parameters and dependencies.
 * refactor: Optimize the runtime of `archive$best()` method with partial sorting.
+* compatibility: Work with new paradox version 1.0.0
 
 # bbotk 0.7.3